refactor(pages): separate search page in components
